+++ src/MicroElements.Metadata/Metadata/Schema/IListMaxItems.cs
@@ -83,7 +83,7 @@ namespace MicroElements.Metadata.Schema
         {
             schema.AssertArgumentNotNull(nameof(schema));
 
-            return schema.GetMetadata<IListMaxItems>();
+            return schema.GetSchemaMetadata<IListMaxItems>();
         }
     }
 }