+++ src/MicroElements.Metadata/Metadata/Schema/IListMinItems.cs
@@ -83,7 +83,7 @@ namespace MicroElements.Metadata.Schema
         {
             schema.AssertArgumentNotNull(nameof(schema));
 
-            return schema.GetMetadata<IListMinItems>();
+            return schema.GetSchemaMetadata<IListMinItems>();
         }
     }
 }