+++ src/MicroElements.Metadata/Metadata/Schema/IStringPattern.cs
@@ -73,7 +73,7 @@ namespace MicroElements.Metadata.Schema
         {
             schema.AssertArgumentNotNull(nameof(schema));
 
-            return schema.GetMetadata<IStringPattern>();
+            return schema.GetSchemaMetadata<IStringPattern>();
         }
 
         /// <summary>