+++ src/MicroElements.Metadata/Metadata/Schema/SimpleTypeSchema.cs
@@ -13,7 +13,7 @@ namespace MicroElements.Metadata.Schema
     /// Example: Currency (string type with maxLength: 3).
     /// </summary>
     [DebuggerTypeProxy(typeof(MetadataProviderDebugView))]
-    public class SimpleTypeSchema : ISchema
+    public class SimpleTypeSchema : ISchema, IManualMetadataProvider
     {
         /// <inheritdoc />
         public string Name { get; }
@@ -38,7 +38,12 @@ namespace MicroElements.Metadata.Schema
             Name = name;
             Type = type;
             Description = description;
+
+            Metadata = new ConcurrentMutablePropertyContainer(searchOptions: MetadataProvider.DefaultSearchOptions);
         }
+
+        /// <inheritdoc />
+        public IPropertyContainer Metadata { get; }
     }
 
     /// <summary>